Instrument ledger latency independently
Currently the ledger latency is monitored as part of the usual
business of writing to the ledger. This means, though, that if
there is a partition to the ctldb, and the reflector is not
able to fetch ledger entries to write, that the ledger latency
will not be instrumented until the partition is resolved.

This change moves the instrumentation behavior to the existing
ledger Monitor which used to only set healthy/unhealthy ECS
attributes.  In this way, it will independently tick and measure
the latency regardless of whether or not the reflector is
applying updates from the ledger.