[F8N-865] Enable globalstats by default (#20)
This PR enables `globalstats` by default. Previously, `globalstats` was opt-in, but this led to unreliable usage/performance stats. Considering we perform batching on these stats, the overhead of enabling `globalstats` is low, so we want to move towards an opt-out strategy instead.

This PR is based on @collinvandyck's proposal here, but takes a lazy initialization approach to avoid launching a goroutine from `init` / simplify the logic for getting the default stats handler: https://github.com/segmentio/ctlstore/compare/init-thoughts?expand=1