[F8N-755] Halt snapshots if supervisor's ledger is out-of-date (#21)
If we were to lose snapshot data in S3, then new reflectors will have to start replaying the full DML ledger from the beginning to recreate an LDB. If this happened, then when the supervisor reboots and starts recreating its LDB, it'll start uploading quite out-of-date snapshots to S3 every 5m. After it uploads its first snapshot, any new reflectors that come online will take that snapshot and start rebuilding their LDB from that point-in-time. That would quickly lead to significant load on the ctldb.

In general, we never want to upload a snapshot unless it is relatively recent. This PR updates the supervisor to inspect its ledger latency when performing a snapshot. The exact limit that it enforces, the maximum ledger latency, is configurable, but will default to `10m`.