Use context.Background() in LDBReader API (#36)
The long story short is that due to how the underlying sqlite
code handles context cancellations, and due to the fact that
we're reading from disk instead of making a network call, using
contexts for this API is potentially more harmful than beneficial.

This change alters the API to use context.Background() instead
of the passed in context to prevent cancellation.  It does this,
instead of removing the context parameter from the exported
APIs, to preserve API compatibility for existing clients.