Use the tx to modify the ctldb. (#52)
We ran into a bug when trying to add columns to an existing
table. The update applied to the ctldb but a ledger entry
was never added. This caused downstream issues as the
LDBs never got the column added.

The cause is that the ctldb change was not using the transaction
handle previously created. This fixes it to use the tx
handle for both the ctldb update and the ledger update.