Apply mysql DDL outside of the current TX.
Even though the statements were re-ordered, applying the DDL last,
even though it failed, still caused the current tx to be committed.

With this change, we apply mysql DDL outside of the current TX to
avoid this behavior. In sqlite, DDL is transactional, so we apply
the statements inside the transaction for sqlite3.

Added regression tests for both CreateTable and AddFields.