Apply DML first, and then DDL (#63)
* Increase the handler timeout to 30s.

This should help prevent some errors when trying to modify
large tables.

* Apply DML, and then DDL. Annotate errors.

Because DDL cannot be rolled back, it makes sense to do this
last. If the DDL fails, the DML can then be rolled back.

* Add additional assertions to the AddFields test.

* Create table applies DML, then DDL.

* Annotate CreateTable errors as well.

* Apply mysql DDL outside of the current TX.

Even though the statements were re-ordered, applying the DDL last,
even though it failed, still caused the current tx to be committed.

With this change, we apply mysql DDL outside of the current TX to
avoid this behavior. In sqlite, DDL is transactional, so we apply
the statements inside the transaction for sqlite3.

Added regression tests for both CreateTable and AddFields.