Pin counterfeiter version. Add optimizations.
Following the README at https://github.com/maxbrunsfeld/counterfeiter
on how to pin dependencies, we now pin counterfeiter to v6.

Additionally, we break out the counterfeiter command into generate.go
files so that the command is only invoked once per package. This is
an optimization and also a way to get rid of an annoying warning.

Changes to the fakes reflect the new version of counterfeiter.