Delete ctlstore website.
I think we will ultimately want to have a website, but it needs
to be more documentation focused, and ideally would have fewer
dependencies that could surface vulnerabilities.

Soon, we will be making pushes to improve the documentation
and onboarding experience for ctlstore.

Before these changes are merged, we will update the DNS
for ctlstore.segment.com to point to our own nginx frontend which
will then in turn redirect to the github repo.