Publish ctlstore docker image to ghcr (#115)
These changes will create two docker images upon merging into master that are pushed to ghcr.io/segmentio/ctlstore repository. 

One is for the linux/amd64 platform and tagged with the $shortSha.
The other is for linux/arm64 and tagged with the $shortSha-arm.

As currently configured, the arm build does not block merging into master because it is much slower due to the action runner not being a native arm platform.

Additionally, PRs will create non-release pushed images that are tagged with the branch name in the tag to allow canary and non-prod releasing.

Lastly, we no longer use segmentio's base alpine image in favor of the open sourced alpine base. 
Double lastly, we updated the chamber dependency.