all: replace segmentio/go-sqlite3 fork with upstream
Our fork does not build on latest version of Alpine, due to a change
in the API that has been fixed upstream but not backported.

The fork only supported two commits. Both commits have been merged
upstream, but one - https://github.com/mattn/go-sqlite3/pull/736 - is
behind a build tag, which is why we need to add a build tag here.